Change default optical property setting; fix uint_fast16_t typo in simulation results; squash compiler warnings

+++ coretrace/simulation_results/simulation_result.hpp
@@ -67,8 +67,8 @@ namespace SolTrace::Result
                                         const SimulationResult &simres);
 
         // Sun results
-        void set_sun_ray_count(uint_fast16_t ray_count) { this->sun_ray_count = ray_count; }
-        uint_fast16_t get_sun_ray_count() { return this->sun_ray_count; }
+        void set_sun_ray_count(uint_fast64_t ray_count) { this->sun_ray_count = ray_count; }
+        uint_fast64_t get_sun_ray_count() { return this->sun_ray_count; }
         void set_sun_dimensions(double width, double height) { this->sun_width = width; this->sun_height = height; }
         void get_sun_dimensions(double& width, double& height) { width = this->sun_width; height = this->sun_height; }
         void set_sun_A_box(double A) { this->A_sun_box = A; }
